API design updates.

+++ design/airtime.go
@@ -28,7 +28,7 @@ var AirtimeValidationPayload = Type("AirtimeValidationPayload", func() {
 		Description("The 3-digist ISO format currency for the value of this transaction")
 		Example("KES")
 	})
-	Attribute("amount", UInt64, func() {
+	Attribute("amount", Float64, func() {
 		Description("Amount - in the provided currency - that the client will receive.")
 		Example(500.00)
 	})
